add toggle to update upstream-sources.yml

+++ SourceStream/file/config.py
@@ -1,13 +1,16 @@
 from ..vendor.SplashPyUtils import config
 
 import requests
+import os
 
 # Loads all config files and download extra config files when needed
 def load():
     YAML_CONFIG = config.configloader("config").load()
 
-    r = requests.get("https://www.enthix.net/SplashOS/downloads/configs/upstream-sources.yml")
-    open("upstream-sources.yml", 'wb').write(r.content)
+
+    if YAML_CONFIG["update-sources"] and not os.path.isfile(".lock-sources"):
+        r = requests.get("https://www.enthix.net/SplashOS/downloads/configs/upstream-sources.yml")
+        open("upstream-sources.yml", 'wb').write(r.content)
     YAML_SOURCES = config.configloader("sources").load()
 
     if YAML_CONFIG["edition"] != "custom":
